# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API",
        "title": "Backend API",
        "contact": {
            "name": "Stanislav Simakhin",
            "url": "https://ssyan.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/oauth/{provider}": {
            "get": {
                "description": "Get redirect URL for the specified OAuth provider (google, yandex, github). Redirects to the provider's authorization page",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get OAuth URL",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "yandex",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Unsupported provider",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/oauth/{provider}/callback": {
            "get": {
                "description": "Handle OAuth provider callback after user authorization. Returns access token in body, sets refresh token as HTTP-only cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "OAuth Callback",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "yandex",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth provider",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OAuth authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Code is required",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error description"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "response.TokenData": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIs..."
                }
            }
        },
        "response.TokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.TokenData"
                },
                "message": {
                    "type": "string",
                    "example": "login successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer [JWT]",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}